Move RefCountProcessor into tests
