+++ src/test/java/io/reactivex/tck/UnicastProcessorTckTest.java
@@ -32,7 +32,7 @@ public class UnicastProcessorTckTest extends IdentityProcessorVerification<Integ
     @Override
     public Processor<Integer, Integer> createIdentityProcessor(int bufferSize) {
         UnicastProcessor<Integer> up = UnicastProcessor.create();
-        return up.refCount();
+        return new RefCountProcessor<Integer>(up);
     }
 
     @Override
